Create Data_Science_PRJT_01.py
Python project that uses data analysis techniques to create graphs from a CSV table of cancellations of a certain product. The project was developed in VSCode using the Jupyter Notebook extension.
Project for demonstration and portfolio purposes only, being freely accessible and free to be downloaded and used as desired. For any questions, contact me on LinkedIn.